+++ src/models/__init__.py
@@ -1,0 +1,6 @@
+
+from .ireview import IReview
+
+__all__ = [
+    "IReview",
+]