+++ CHANGELOG.md
@@ -7,5 +7,6 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Created `IDatabase` class in `src/data/idatabase.py`. [dev/aingeru]
 - Created `SQLiteDatabase` class in `src/data/sqllitedatabase.py`. [dev/aingeru]
+- Created `IReview` class in `src/models/ireview.py`. [dev/aingeru]
 
 ### Changed